third_party/blink/web_tests/external/wpt/fledge/tentative/kanon-status-below-threshold.https.window-expected.txt

This is a testharness.js-based test.
[FAIL] Check kAnonStatus is "belowThreshold" when FledgeConsiderKAnonymityis enabled and FledgeEnforceKAnonymity is disabled
  assert_array_equals: expected property 0 to be "https://web-platform.test:8444/fledge/tentative/resources/request-tracker.py?uuid=<uuid>&dispatch=track_get&id=bidder_report_1" but got "https://web-platform.test:8444/fledge/tentative/resources/request-tracker.py?uuid=<uuid>&dispatch=track_get&id=bidder_report_error" (expected array ["https://web-platform.test:8444/fledge/tentative/resources/request-tracker.py?uuid=<uuid>&dispatch=track_get&id=bidder_report_1"] got ["https://web-platform.test:8444/fledge/tentative/resources/request-tracker.py?uuid=<uuid>&dispatch=track_get&id=bidder_report_error"])
Harness: the test ran to completion.
